requirements.txt
alembic==1.0.7
blinker==1.4
Click==7.0
dominate==2.3.1
Flask==1.0.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.3.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-Uploads==0.2.1
Flask-WTF==0.14.2
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.1.0
pkg-resources==0.0.0
psycopg2==2.7.5
psycopg2-binary==2.7.7
PyJWT==1.6.4
python-dateutil==2.8.0
python-editor==1.0.4
six==1.12.0
SQLAlchemy==1.2.17
visitor==0.1.3
Werkzeug==0.14.1
WTForms==2.2.1
